src/requirements.txt
asgiref==3.4.1
attrs==21.2.0
boto3==1.19.12
botocore==1.22.12
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-phonenumber-field==5.2.0
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
Faker==9.8.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.2
phonenumberslite==8.12.36
Pillow==8.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
